Move BackupTimer to external file and rename to Scheduler
